Move cryptography tests to test_cryptography.py and add some more
