+++ more/jwtauth/tests/test_jwtauth.py
@@ -9,20 +9,9 @@ from webtest import TestApp as Client
 from jwt import InvalidIssuerError
 import pytest
 
-try:
-    from cryptography.hazmat.primitives.asymmetric import ec  # noqa
-    has_crypto = True
-except ImportError:
-    has_crypto = False
-
 
 def setup_module(module):
     morepath.disable_implicit()
-
-
-def relative(filepath):
-    import os
-    return os.path.join(os.path.dirname(__file__), filepath)
 
 
 def test_jwt_custom_settings():
@@ -91,54 +80,6 @@ def test_encode_decode_with_issuer():
         'iss': 'Issuer_App'
     }
     claims_set = identity_policy.create_claims_set(userid, extra_claims)
-    token = identity_policy.encode_jwt(claims_set)
-    claims_set_decoded = identity_policy.decode_jwt(token)
-
-    assert claims_set_decoded == claims_set
-
-
-@pytest.mark.skipif(not has_crypto, reason='Not supported without cryptography library')
-def test_encode_decode_with_es256():
-    identity_policy = JWTIdentityPolicy(
-        algorithm='ES256',
-        private_key_file=relative('keys/testkey_ec'),
-        public_key_file=relative('keys/testkey_ec.pub')
-    )
-    claims_set = {
-        'sub': 'user'
-    }
-    token = identity_policy.encode_jwt(claims_set)
-    claims_set_decoded = identity_policy.decode_jwt(token)
-
-    assert claims_set_decoded == claims_set
-
-
-@pytest.mark.skipif(not has_crypto, reason='Not supported without cryptography library')
-def test_encode_decode_with_ps384():
-    identity_policy = JWTIdentityPolicy(
-        algorithm='PS384',
-        private_key_file=relative('keys/testkey_rsa'),
-        public_key_file=relative('keys/testkey_rsa.pub')
-    )
-    claims_set = {
-        'sub': 'user'
-    }
-    token = identity_policy.encode_jwt(claims_set)
-    claims_set_decoded = identity_policy.decode_jwt(token)
-
-    assert claims_set_decoded == claims_set
-
-
-@pytest.mark.skipif(not has_crypto, reason='Not supported without cryptography library')
-def test_encode_decode_with_rs512():
-    identity_policy = JWTIdentityPolicy(
-        algorithm='RS512',
-        private_key_file=relative('keys/testkey_rsa'),
-        public_key_file=relative('keys/testkey_rsa.pub')
-    )
-    claims_set = {
-        'sub': 'user'
-    }
     token = identity_policy.encode_jwt(claims_set)
     claims_set_decoded = identity_policy.decode_jwt(token)
